ai-service/requirements.txt
# AI Service — Smart Garden
# Python 3.10+

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1

# ML & Data Science
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
xgboost>=2.0
optuna>=3.6
joblib>=1.4
matplotlib>=3.8

# Plant Image Analysis (YOLO)
ultralytics>=8.4.0
Pillow>=10.0
opencv-python>=4.8.0